CSAT-English/requirements.txt
torch
transformers
datasets
accelerate
peft
bitsandbytes
numpy
pandas
tqdm
python-docx
scikit-learn